feat(java): detect more java on different drives
