+++ src-tauri/Cargo.toml
@@ -81,6 +81,7 @@ reqwest-retry = "0.7.0"
 murmur2 = "0.1"
 csv = "1.3"
 semver = "1"
+sysinfo = "0.36.0"
 
 [target."cfg(windows)".dependencies]
 winreg = "0.55.0"
