+++ src-tauri/src/launch/helpers/command_generator.rs
@@ -8,8 +8,8 @@ use crate::instance::models::misc::{InstanceError, InstanceSubdirType};
 use crate::launch::helpers::file_validator::get_nonnative_library_paths;
 use crate::launch::helpers::misc::{get_separator, replace_arguments};
 use crate::launch::models::{LaunchError, LaunchingState};
-use crate::launcher_config::helpers::memory::get_memory_info;
 use crate::launcher_config::models::*;
+use crate::utils::sys_info::get_memory_info;
 use base64::engine::general_purpose;
 use base64::Engine;
 use serde::{self, Deserialize, Serialize};
