+++ src-tauri/src/launcher_config/helpers/mod.rs
@@ -1,4 +1,3 @@
 pub mod java;
-pub mod memory;
 pub mod misc;
 pub mod updater;
