+++ src-tauri/src/utils/commands.rs
@@ -1,7 +1,7 @@
 use crate::error::SJMCLResult;
-use crate::launcher_config::helpers::memory::get_memory_info;
 use crate::launcher_config::models::{LauncherConfigError, MemoryInfo};
 use crate::utils::fs::extract_filename as extract_filename_helper;
+use crate::utils::sys_info::get_memory_info;
 use font_loader::system_fonts;
 use std::fs;
 use tauri_plugin_http::reqwest;
